remove edns padding in enhanced dns server response
