+++ clash-netstack/Cargo.toml
@@ -7,6 +7,9 @@ edition.workspace = true
 authors.workspace = true
 homepage.workspace = true
 
+[features]
+default = []
+
 [dependencies]
 smoltcp = { version = "0.12", default-features = false, features = [
     "std",
@@ -20,16 +23,16 @@ smoltcp = { version = "0.12", default-features = false, features = [
     "async"
 ]}
 bytes = "1.0"
-thiserror = "1.0"
 tokio = { version = "1.0", features = ["full"] }
 futures = "0.3"
 log = "0.4"
 rand = "0.9"
 etherparse = "0.16"
 
+
 [dev-dependencies]
 tun-rs = { version = "2", features = ["async", "async_framed"] }
 ipnet = "2"
-env_logger = "0.11"
 socket2 = "0.5.6"
 libc = "0.2"
+env_logger = "0.11"
