+++ clash-netstack/README.md
@@ -2,6 +2,15 @@
 
 See the [examples](examples) directory for how to use the `clash-netstack` library.
 
+
+
+## Run Tests
+To run the tests, use the following command:
+
+```bash
+cargo test -p watfaq-netstack
+```
+
 ## Disclaimer
 
 Heavily based on [netstack-smoltcp](https://github.com/automesh-network/netstack-smoltcp) with simplified API and some modifications to fit the `clash-rs` ecosystem.
